add permission • end

+++ app/Http/Controllers/BookController.php
@@ -18,20 +18,23 @@ class BookController extends Controller
     }
 
     public function create() {
+        $this->authorize('create', Book::class);
+
         return view('books.create', [
             'categories' => Category::all(),
             'authors' => Author::all(),
         ]);
     }
 
-    public function store(Request $request) {
+    public function store(Request $request, Book $book) {
+        $this->authorize('create', $book);
 
         $request->validate([
             'title' => 'required',
             'description' => 'required',
             'nbre_pages' => 'required',
             'category_id' => 'required',
-            'category_id' => 'required',
+            'author_id' => 'required',
             'book_path' => 'required',
         ]);
 
@@ -55,11 +58,39 @@ class BookController extends Controller
         return redirect()->route('books.index');
     }
 
-    public function edit(Book $book, Category $category) {
+    public function edit(Book $book) {
+        $this->authorize('update', $book);
+
         return view('books.edit', compact('book'));
     }
 
+    public function update(Request $request, Book $book) {
+        
+        $request->validate([
+            'title' => 'required',
+            'description' => 'required',
+            'nbre_pages' => 'required',
+            'category_id' => 'required',
+            'author_id' => 'required',
+            'book_path' => 'required',
+        ]);
+
+        $filename = time() . '.' . $request->book_path->extension();
+
+        $book_path = $request->file('book_path')->storeAs(
+            'books_pdf',
+            $filename,
+            'public'
+        );
+
+        $book->update($request->$book_path);
+        $book->update($request->except(['book_path']));
+        return redirect(route('books.index'));
+    }
+
     public function destroy(Book $book) {
+        $this->authorize('delete', $book);
+
         $book->delete();
         return redirect()->route('books.index');
     }
